+++ src/com/intellij/idea/plugin/hybris/system/type/meta/TSMetaModelAccess.kt
@@ -47,7 +47,7 @@ interface TSMetaModelAccess {
     fun findMetaForDom(dom: EnumType): TSGlobalMetaEnum?
     fun <T : TSGlobalMetaClassifier<*>> getAll(metaType: TSMetaType): Collection<T>
     fun getAll(): Collection<TSGlobalMetaClassifier<*>>
-    fun getNextAvailableTypeCode(): Int
+    fun getNextAvailableTypeCode(): Int?
     fun getReservedTypeCodes(): Map<Int, String>
     fun getAllOf(vararg metaTypes: TSMetaType): Collection<TSGlobalMetaClassifier<*>>
 }