+++ src/com/intellij/idea/plugin/hybris/system/type/model/Deployment.java
@@ -1,6 +1,7 @@
 /*
- * This file is part of "hybris integration" plugin for Intellij IDEA.
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
  * Copyright (C) 2014-2016 Alexander Bartash <AlexanderBartash@gmail.com>
+ * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -21,6 +22,8 @@
 
 package com.intellij.idea.plugin.hybris.system.type.model;
 
+import com.intellij.idea.plugin.hybris.system.type.util.xml.converter.DeploymentTypeCodeConverter;
+import com.intellij.util.xml.Convert;
 import com.intellij.util.xml.DomElement;
 import com.intellij.util.xml.GenericAttributeValue;
 import com.intellij.util.xml.Required;
@@ -66,6 +69,7 @@ public interface Deployment extends DomElement {
     @NotNull
     @com.intellij.util.xml.Attribute(TYPE_CODE)
     @Required
+    @Convert(DeploymentTypeCodeConverter.class)
     GenericAttributeValue<String> getTypeCode();